added improvements to promptedto_cons
I've improved some of the prompting to get better results.  however, i've found that the biggest improvement came from running it in sync vs async (for neg checker errors reduced by 1/2, 0.04 compared to 0.08).  I'm not sure if this is the reason, but maybe there's some type of load throttling with api.